Adding more commands to cli support
