+++ examples/transaction.txt
@@ -1,3 +1,3 @@
 select A0000000043060
 get_processing_options
-generate_ac 80 0000000000100000000000100986000000000009861504280030901B6A2300001EABC126F85499760000000000000000000000000000000000000000000000000000
+generate_ac ARQC 0000000000100000000000100986000000000009861504280030901B6A2300001EABC126F85499760000000000000000000000000000000000000000000000000000
