+++ src/lib.rs
@@ -1,5 +1,5 @@
 use capdu::APDU;
-use capdu::CryptogramType;
+pub use cli::{Command, Emv, Mode};
 pub use connection::connect;
 use rapdu::{RAPDU, Status};
 use tlv::TLV;
@@ -10,9 +10,36 @@ mod rapdu;
 mod connection;
 mod utils;
 mod banner;
+mod cli;
 
 pub const MASTERCARD_MAESTRO: [u8; 7] = [0xA0, 0x00, 0x00, 0x00, 0x04, 0x30, 0x60];
 pub const MASTERCARD_CREDIT: [u8; 7] = [0xA0, 0x00, 0x00, 0x00, 0x04, 0x10, 0x10];
+
+#[derive(Debug)]
+pub enum CryptogramType {
+    AAC,
+    ARQC,
+    TC,
+}
+
+impl CryptogramType {
+    pub fn to_reference_control(&self) -> u8 {
+        match self {
+            CryptogramType::AAC => 0x00,
+            CryptogramType::ARQC => 0x80,
+            CryptogramType::TC => 0x40
+        }
+    }
+
+    pub fn from_str(str: &str) -> CryptogramType {
+        match str {
+            "AAC" => CryptogramType::AAC,
+            "ARQC" => CryptogramType::ARQC,
+            "TC" => CryptogramType::TC,
+            _ => panic!("Unknown cryptogram type")
+        }
+    }
+}
 
 fn send(card: &pcsc::Card, apdu: APDU) {
     connection::transmit(card, &apdu)
